requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
django-filter==24.3
django-jazzmin==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
packaging==24.1
PyJWT==2.9.0
pyotp==2.9.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
shortuuid==1.0.13
simplejson==3.19.3
sqlparse==0.5.1
twilio==9.2.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
